pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mozaik-generator"
version = "0.1.0"
description = "Generate photomosaic images from tile image libraries"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
mozaik = "mozaik.cli:main"

[tool.setuptools.packages.find]
include = ["mozaik*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
